requirements.txt
flask
pyyaml
PyPDF2
google-generativeai
Flask-BasicAuth